requirements.txt
requests==2.27.1
psycopg2-binary==2.9.3
pydantic==1.9.1
